remade structure3

+++ src/sample/Drone.java
@@ -3,13 +3,13 @@ package sample;
 import javafx.scene.canvas.GraphicsContext;
 
 public class Drone {
-    int x;
-    int y;
+    double x;
+    double y;
     int length = 20;
     int width = 20;
     private boolean active = false;
 
-    public Drone(int x, int y) {
+    public Drone(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -19,19 +19,19 @@ public class Drone {
 
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
